Dev obr

+++ src/backend/classes/Files.py
@@ -443,6 +443,8 @@ class Files:
         rand = ''.join(random.choice(string.ascii_lowercase) for i in range(4))
         filename = filename.replace(' ', '_') + '_' + rand + file_ext.lower()
         new_path = os.path.join(path, secure_filename(filename))
+        print('new_path : ')
+        print(new_path)
         file.save(new_path)
         return new_path
 
@@ -523,9 +525,11 @@ class Files:
     def zip_files(input_paths, output_path, delete_zipped_files=False):
         with ZipFile(output_path, 'w') as zipObj:
             for input_path in input_paths:
-                zipObj.write(input_path['input_path'], input_path['path_in_zip'])
-                if delete_zipped_files:
-                    os.remove(input_path['input_path'])
+                if os.path.exists(input_path['input_path']):
+                    zipObj.write(input_path['input_path'],
+                                 input_path['path_in_zip'] if input_path['path_in_zip'] else None)
+                    if delete_zipped_files:
+                        os.remove(input_path['input_path'])
 
 
 def compress_pdf(input_file, output_file, compress_id):
